+++ app/src/main/java/com/omtorney/doer/core/data/database/NoteDao.kt
@@ -1,7 +1,7 @@
 package com.omtorney.doer.core.data.database
 
 import androidx.room.*
-import com.omtorney.doer.core.model.Note
+import com.omtorney.doer.notes.domain.model.Note
 import kotlinx.coroutines.flow.Flow
 
 @Dao
@@ -11,7 +11,7 @@ interface NoteDao {
     fun getNotes(): Flow<List<Note>>
 
     @Query("SELECT * FROM notes WHERE id = :id")
-    suspend fun getNoteById(id: Int): Note?
+    suspend fun getNoteById(id: Long): Note?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(note: Note)
